refactor(scripts): Phase 3 file-size cleanup + docstring update
code-reviewer MEDIUM 指摘対応 — align.py / extract.py が 800 行 soft cap を
超えたため責務分離で解消:

**align.py 852 行 → 733 行**:
- ALIGN_OUTPUT_SCHEMA_VERSION + 5 つの ``build_diff_*`` factory +
  ``section_label`` を ``align_diffs.py`` (171 行) に分離
- align.py 本体は public API (``align_segments`` / ``parity_diffs_to_issues``) +
  section 分割 + weighted LCS + cross-section shift 検出に集中
- diff factory は ``align.py`` から import alias 経由で使う (``build_diff_*``)

**extract.py 906 行 → 642 行**:
- line-level state machine (``classify_line`` / section 別 step/bullet/paragraph
  count) を ``extract_lines.py`` (314 行) に分離
- ``extract.py`` 末尾で ``from .extract_lines import ...`` して
  backward-compatible な surface を維持 — consumer (checks.py / tests /
  harness) の import path は変わらない
- extract.py 本体は markdown atom primitives (image / callout / heading /
  strip / untranslated / table) と invariant token に集中

**__init__.py docstring 更新** (code-reviewer MEDIUM):
- Phase 3 の 12 module (M1-M4) + Phase 1-2 の segment extractors を
  layer 順で明記 (基盤 → extractors → M1 predicates → M2 supporting →
  M3/M4 比較エンジン)
- M5-M7 が別 PR で追加予定である旨を note

Verification:
- 682 tests pass (unit + conformance, 288-matrix を除く)
- 92.62% coverage (threshold 90%)
- ruff / ruff format / mypy clean (30 source files)
- 行動変更なし — 純粋な責務分離と re-export 追加のみ

288-page matrix の再実行は local で timeout したため本 commit では未検証。
harness と production code 双方未変更で、diff factory と line-level state
machine は pure function のため重回帰リスクは低い。CI の matrix test が
最終検証を担う。

Refs: Issue #368 Phase 3 reviewer gate (code-reviewer MEDIUM 3 件 解消)
